backend/requirements.txt
# FastAPI and web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# LangChain and AI
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Vector Database
chromadb>=0.5.0

# Document Processing
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# CORS and middleware
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
